updated setup.sh; added doom config

+++ setup.sh
@@ -3,13 +3,26 @@
 GREEN="\x1b[32m"
 NC="\x1b[0m"
 
+remove=false
+is_test=false
+
+for arg in "$@"; do
+	case "$arg" in
+		-t|--t|-test|--test) is_test=true ;;
+		-r|--r|-remove|--remove) remove=true ;;
+		*) echo "unknown flag: $arg"; exit 1 ;;
+	esac
+done
+
 add_line_to_zshrc_if_not_exists() {
     line="$1"
     if [[ ! $(cat "$HOME/.zshrc" | grep "$line" >> /dev/null; echo $?) -eq 0 ]]; then
         echo "Adding line to .zshrc: $line"
         zshrc="$HOME/.zshrc"
         temp="$DOTFILES/temp-zshrc"
-        { echo -e "$line\n"; cat "$zshrc"; } > "$temp" && mv "$temp" "$zshrc"
+		if ! $is_test; then
+        	{ echo -e "$line\n"; cat "$zshrc"; } > "$temp" && mv "$temp" "$zshrc"
+		fi
     fi
 }
 
@@ -29,23 +42,25 @@ update_directory() {
 		fi
 
 		printf "${GREEN}rsync -ia --delete \"$dir\" \"$to_dir\"$NC\n"
-		rsync -ia --delete "$dir" "$to_dir"
+		if ! $is_test; then
+			rsync -ia --delete "$dir" "$to_dir"
+		fi
 	done
 }
 
-if [[ "$OSTYPE" == "darwin"* ]]; then
-	# dock
-	defaults write com.apple.dock autohide -bool true
-	defaults write com.apple.dock mru-spaces -bool true
-	# finder
-	defaults write com.apple.finder AppleShowAllExtensions -bool true
-	defaults write com.apple.finder FXPreferredViewStyle -string "clmv"
-	defaults write com.apple.finder FXRemoveOldTrashItems -bool true
-	# disable Ctrl + Space, because of tmux prefix
-	defaults write com.apple.symbolichotkeys AppleSymbolicHotKeys -dict-add 60 "{ enabled = 0; value = { parameters = (32, 49, 262144); type = standard; }; }"
+if ! $is_test; then
+	if [[ "$OSTYPE" == "darwin"* ]]; then
+		# dock
+		defaults write com.apple.dock autohide -bool true
+		defaults write com.apple.dock mru-spaces -bool true
+		# finder
+		defaults write com.apple.finder AppleShowAllExtensions -bool true
+		defaults write com.apple.finder FXPreferredViewStyle -string "clmv"
+		defaults write com.apple.finder FXRemoveOldTrashItems -bool true
+		# disable Ctrl + Space, because of tmux prefix
+		defaults write com.apple.symbolichotkeys AppleSymbolicHotKeys -dict-add 60 "{ enabled = 0; value = { parameters = (32, 49, 262144); type = standard; }; }"
+	fi
 fi
-
-arg=$1
 
 dotfiles=$(realpath "$0" | sed 's|\(.*\)/.*|\1|')
 
@@ -53,22 +68,35 @@ export DOTFILES="$dotfiles"
 
 config_bf_dotfiles="$HOME/.config-bf-dotfiles"
 zshrc_bf_dotfiles="$HOME/.zshrc-bf-dotfiles"
+doomd_bf_dotfiles="$HOME/.doom.d"
 
-if [[ "$arg" == "-r" ]]; then
-    read -p "Are you shure you want to remove the dotfiles and get your old config back? [y/N]: " continue
+if $remove; then
+    read -p "Are you shure you want to remove the dotfiles and hopefully get your old config back? [y/N]: " continue
     continue=${continue:-n}
 
     if [[ "$continue" == "y" ]]; then
         if [[ -d "$config_bf_dotfiles" ]]; then
             echo "Resetting .config..."
             update_directory "$config_bf_dotfiles" "$HOME/.config"
-            rm -rf "$config_bf_dotfiles"
+			if ! $is_test; then
+            	rm -rf "$config_bf_dotfiles"
+			fi 
         fi
 
         if [[ -f "$zshrc_bf_dotfiles" ]]; then
             echo "Resetting .zshrc..."
-            cp "$zshrc_bf_dotfiles" "$HOME/.zshrc"
-            rm "$zshrc_bf_dotfiles"
+			if ! $is_test; then
+				cp "$zshrc_bf_dotfiles" "$HOME/.zshrc"
+				rm "$zshrc_bf_dotfiles"
+			fi
+        fi
+
+        if [[ -d "$doomd_bf_dotfiles" ]]; then
+            echo "Resetting .doom.d..."
+            update_directory "$doomd_bf_dotfiles" "$HOME/.doom.d"
+			if ! $is_test; then
+            	rm -rf "$doomd_bf_dotfiles"
+			fi
         fi
 
         echo "Make sure to delte the $dotfiles direcotry."
@@ -80,10 +108,17 @@ else
         echo "Making copy of $HOME/.config to $config_bf_dotfiles"
         update_directory "$HOME/.config" "$config_bf_dotfiles" 1
     fi
-    
+
     if [[ ! -f "$zshrc_bf_dotfiles" ]]; then
         echo "Making copy of $HOME/.zshrc to $zshrc_bf_dotfiles"
-        cp "$HOME/.zshrc" "$zshrc_bf_dotfiles"
+		if ! $is_test; then
+        	cp "$HOME/.zshrc" "$zshrc_bf_dotfiles"
+		fi
+    fi
+
+    if [[ ! -d "$doomd_bf_dotfiles" ]]; then
+        echo "Making copy of $HOME/.doom.d to $doomd_bf_dotfiles"
+        update_directory "$HOME/.doom.d" "$config_bf_dotfiles" 1
     fi
 
     for file in $(find $dotfiles -name 'map'); do
@@ -104,7 +139,9 @@ else
                 	link=$(eval echo "$link")
 					from_to=${link:3}
 					printf "${GREEN}cp $from_to$NC\n"
-					eval "cp $from_to"
+					if ! $is_test; then
+						eval "cp $from_to"
+					fi
 				fi
 			fi
         done <$file
